% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dje.R
\name{dje}
\alias{dje}
\title{dje: Test for Differential Junction Expression}
\usage{
dje(fit, geneid, junctionID = NULL, robust = FALSE, verbose = TRUE)
}
\arguments{
\item{fit}{an MArrayLM fitted model object produced by lmFit or contrasts.fit. Rows should correspond to junctions.}

\item{geneid}{gene identifiers. Either a vector of length nrow(fit) or the name of the column of fit$genes containing the gene identifiers. Rows with the same ID are assumed to belong to the same gene.}

\item{junctionID}{junction identifiers. Either a vector of length nrow(fit) or the name of the column of fit$genes containing the junction identifiers.}

\item{robust}{logical, should the estimation of the empirical Bayes prior parameters be robustified against outlier sample variances?}

\item{verbose}{logical, if TRUE some diagnostic information about the number of genes and junctions is output.}
}
\description{
Given a linear model fit at the junction level, test for differences in junction expression between experimental conditions. This is a wrapper function of limma diffSplice.
}
\section{Details}{

An object of class MArrayLM containing both junction level and gene level tests. Results are sorted by geneid and by junctionid within gene.
See limma::diffSplice documentation for details.
}

\section{Value}{

This function tests for differential junction usage for each gene and for each column of fit following limma diffSplice implementation.
}

\examples{
\dontrun{
fit.norm0 <- system.file("extdata", "fit.norm.rds", package = "DJExpress")
fit.norm <- readRDS(fit.norm0)
groupID0 <- system.file("extdata", "groupID.rds", package = "DJExpress")
groupID <- readRDS(groupID0)
junctionID0 <- system.file("extdata", "junctionID.rds", package = "DJExpress")
junctionID <- readRDS(junctionID0)
dje.out <- DJExpress::dje(fit = fit.norm, geneid=groupID, junctionID=junctionID)
}
}
