% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEanalyze.R
\name{DJEanalyze}
\alias{DJEanalyze}
\title{DJEanalyze: Test for Differential Junction Usage}
\usage{
DJEanalyze(
  prepare.out,
  Group1,
  junct.annot = NULL,
  normalize.method = "none",
  calcNormFactors = FALSE,
  calMethod = "none",
  plot = FALSE,
  FDR = 0.05,
  logFC = 2,
  level = 0.9,
  topJunct = NULL,
  targetGene = NULL,
  legend.position = "right",
  volc.axes.face = "plain",
  volc.axes.col = "black",
  volc.axes.size = 11,
  volc.axes.angle = 0,
  scale_color_manual = c("#005BA2", "#D44F4F", "#B3B4B5")
)
}
\arguments{
\item{prepare.out}{output object from DJEprepare()}

\item{Group1}{vector or factor specifying basic control sample names}

\item{junct.annot}{gene-annotated junction IDs}

\item{normalize.method}{limma voom's normalization method to be applied to the logCPM values (default "none")}

\item{calcNormFactors}{logical, should edgeR's library Size Normalization be calculated?}

\item{calMethod}{Normalization method to be used. Options are: "TMM","RLE","upperquartile" or "none" (default "none")}

\item{plot}{logical, should median/variance trend plot be displayed?}

\item{FDR}{numeric: adjusted p-value cutoff in the linear model fit to define significance}

\item{logFC}{numeric: log2-fold-change value cutoff corresponding to the effect or contrast in the linear model fit to define significance}

\item{level}{numeric: confidence interval for Linear Regression of absolute vs gene-wise logFC in junction expression}

\item{topJunct}{numeric: number of top significant junctions that should be labeled in volcano plot}

\item{targetGene}{vector or factor specifying the genes whose significant junctions should be labeled in volcano plot}

\item{legend.position}{character indicating the position of labels in volcano plot. Allowed values are: “left”, “top”, “right”, “bottom”. Default "right".}

\item{volc.axes.face}{character indicating the font face of axes tick labels in volcano plot. Allowed values are: "plain", "italic", "bold", "bold.italic"}

\item{volc.axes.col}{character indicating the color of axes tick labels in volcano plot}

\item{volc.axes.size}{numeric indicating the size of axes tick labels in volcano plot}

\item{volc.axes.angle}{numeric indicating the angle of axes tick labels in volcano plot}

\item{scale_color_manual}{vector of aesthetic values (colors) to indicate significance of data values in volcano plot (default are #005BA2 for downregulated, #D44F4F for upregulated and #B3B4B5 for non-significant junctions)}
}
\value{
A list of objects containing:
v.norm: log-cpm values output by limma:voom
ex.norm: Differential Junction Expression analysis output
dje.out: Annotated ex.norm data set with additional information, including basic statistics (e.g. median, zero counts, etc) and DJE group for each junction.
dje.sig: Significant hits in dje.out based on FDR and logFC cutoffs.
volcano.plot: Volcano plot of differential junction expression
logFC.plot.junctions": Table with junctions shown in logFC.plot. They are defined as junctions passing FDR cutoff for differential usage as well as differential expression.
logFC.plot: Regression plot of Absolute logFC ~ Relative logFC for differentially used (compared to average junction expression in the gene) and differentially expressed junctions (basal vs tested sample group)
}
\description{
Normalize junction expression and performs differential junction usage analysis using limma::diffSplice methods. Returns output object for DJEplotSplice(), DJEvsTrait() and JCNAprepare() functions.
}
\section{Details}{

This wrapper function receives DJEprepare() output object and performs junction expression normalization and differential junction usage analysis. The differential analysis is an implementation of limma::diffSplice() method.
See limma::diffSplice documentation for details.
}

\examples{
DJEprep <- system.file("extdata", "DJEprep.rds", package = "DJExpress")
prep.out <- readRDS(DJEprep)
Group1 <- colnames(prep.out$JunctExprfilt)[grep("GTEx", colnames(prep.out$JunctExprfilt))]
anlz.out <- DJExpress::DJEanalyze(prep.out, Group1)
}
\seealso{
\code{\link{limma}}
}
