% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEanlz.R
\docType{data}
\name{DJEanlz}
\alias{DJEanlz}
\title{output object of DJEanalize function}
\format{
An object of class \code{"list"}.
}
\usage{
data(DJEanlz)
}
\description{
A list of objects containing the results of differential junction expression analysis. TCGA and GTEx example junction expression is used.
}
\examples{
data(DJEanlz)
iPlot.out <- DJEplotSplice(DJEanlz, geneID="RPSA")
}
\references{
Kahles et al.Cancer Cell (2018) Aug 13;34(2):211-224.e6.
(\href{https://pubmed.ncbi.nlm.nih.gov/30078747/}{PubMed})
}
\keyword{datasets}
