% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEimport.R
\name{DJEimport}
\alias{DJEimport}
\title{DJEimport: Construction of junction expression matrix}
\usage{
DJEimport(
  workDir = NULL,
  data.type = c("sample", "matrix"),
  aligner = c("STAR", "other"),
  min.expressed = 3
)
}
\arguments{
\item{workDir}{path to folder where individual junction quantification files (e.g. from STAR alignment) are located. This folder should only contain junction quantification files.}

\item{data.type}{One of c("sample", "matrix"), indicating whether the input files are matrices per individual samples or joint into a single expression matrix.}

\item{aligner}{One of c("STAR", "other"), indicating the alignment tool used to produce junction quantification. Only used if data.type = "matrix".
If "other" is indicated, files in workDir path should contain junction IDs (with the format chr:start:end:strand) in the first column read counts in the second column.}

\item{min.expressed}{Numeric. Minimal number of junction reads to be considered expressed in sample. Default is 3 reads.}
}
\value{
Matrix of samples as columns and junction quantifications as rows and associated matrix with genomic coordinates.
}
\description{
Uses sample-level junction quantification files to construct expression matrix. Returns output object for DJEannotate.
}
\examples{
in.file <- system.file("extdata", "junct.quant", package = "DJExpress")
out.file <- DJEimport(in.file, aligner="STAR")
}
