% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEplotSplice.R
\name{DJEplotSplice}
\alias{DJEplotSplice}
\title{DJEplotSplice: Differential junction usage plot}
\usage{
DJEplotSplice(
  analize.out,
  geneID = NULL,
  gtf = NULL,
  target.junction = NULL,
  genecolname = NULL,
  rank = 1L,
  logFC = 2,
  meanExp = 10,
  FDR = 0.05
)
}
\arguments{
\item{analize.out}{output object from DJEanalyze()}

\item{geneID}{gene name for which plot is generated}

\item{gtf}{Reference transcriptome in genecode gtf file format. Used to define gene coordinates}

\item{target.junction}{Junction IDs to highlight within gene model plot. Maximum 5 junctions.}

\item{genecolname}{column name of fit$genes containing gene IDs. Defaults to fit$genecolname}

\item{rank}{integer, if geneid=NULL then this ranked gene will be plotted}

\item{logFC}{numeric: log2-fold-change value cutoff in junction usage to define significance in plot}

\item{meanExp}{numeric: mean expression value cutoff in tested condition group to define significance in plot}

\item{FDR}{numeric: adjusted p-value cutoff in the linear model fit to define significance}
}
\value{
interactive highcharts plot showing relative logFC across junctions in target gene
}
\description{
Plots relative log-fold changes by junctions for the specified gene and highlights the significantly spliced junctions. This is a wrapper function of limma plotSplice.
}
\examples{
data(DJEanlz)
iPlot.out <- DJEplotSplice(DJEanlz, geneID="ENAH", logFC = 0.8, FDR = 0.05)
}
