% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEprepare.R
\name{DJEprepare}
\alias{DJEprepare}
\title{DJEprepare: Filtering junctions for differential expression analysis}
\usage{
DJEprepare(
  annotate.out,
  Group1,
  minMean = 10,
  maxMean = Inf,
  minVar = 0,
  maxVar = Inf
)
}
\arguments{
\item{annotate.out}{output object from DJEannotate()}

\item{Group1}{vector or factor specifying sample names in the control group}

\item{minMean}{numeric: minimum of read count mean per junction}

\item{maxMean}{numeric: maximum of read count mean per junction}

\item{minVar}{numeric: minimum of read count variance per junction}

\item{maxVar}{numeric: maximum of read count variance per junction}
}
\value{
List object containing expression-based filtered junction counts, gene annotation and design matrix for differential expression analysis
}
\description{
Filter junctions based on their expression. Returns output object for DJEanalize
}
\examples{
DJEann <- system.file("extdata", "DJEann.rds", package = "DJExpress")
ann.out <- readRDS(DJEann)
Group1 <- colnames(ann.out$quant.annotated)[grep("GTEx", colnames(ann.out$quant.annotated))]
prep.out <- DJEprepare(ann.out, Group1)
}
