% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEspliceRadar.R
\name{DJEspliceRadar}
\alias{DJEspliceRadar}
\title{DJEspliceRadar: Radar plot of junction expression vs trait correlation}
\usage{
DJEspliceRadar(
  DJEvsTrait.out,
  base.junction = NULL,
  topTraits = NULL,
  select.traits = NULL,
  ordered.junction = NULL
)
}
\arguments{
\item{DJEvsTrait.out}{output object from DJEvsTrait()}

\item{base.junction}{vector of junction IDs to be used for trait association}

\item{topTraits}{numeric: number of top junction-associated traits to be shown in the plot}

\item{select.traits}{vector of trait names to be shown in the plot}

\item{ordered.junction}{junction ID for which correlation coefficient values are sorted in descending order in the plot}
}
\value{
Interactive highcharts-based radar plot showing top associations between selected junction expression and traits based on correlation coefficients
}
\description{
Interactive plot with top correlated junction expression vs sample traits obtained with DJEvsTrait function.
}
\examples{
DvT <- system.file("extdata", "DvT.rds", package = "DJExpress")
DvT.out <- readRDS(DvT)
Sr.out <- DJEspliceRadar(DvT.out, ordered.junction = "chr1:225688773:225692692:2")
}
