% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DJEvsTrait.R
\name{DJEvsTrait}
\alias{DJEvsTrait}
\title{DJEvsTrait: Junction relationship to external trait}
\usage{
DJEvsTrait(
  analize.out,
  Group1,
  traitData,
  nThreads = 2,
  select.junctions = NULL,
  coeff = 0.3,
  pvalue = 0.05,
  plot.heatmap = FALSE,
  test.type = c("Correlation", "Regression", "both"),
  cor.method = c("bicor", "pearson", "kendall", "spearman"),
  useModel = c(modelANOVA, modelLINEAR, modelLINEAR_CROSS),
  covariates = NULL
)
}
\arguments{
\item{analize.out}{output object from DJEanalize()}

\item{Group1}{vector or factor specifying basic control sample names}

\item{traitData}{a numeric vector or a matrix of external sample traits}

\item{nThreads}{numeric: number of threads to allow.}

\item{select.junctions}{a vector of junction IDs used to generate SpliceRadar plots when test.type = "Correlation". Junction ID vector length should be < 5}

\item{coeff}{numeric: correlation coefficient cutoff for significance}

\item{pvalue}{numeric: correlation p-value cutoff for significance}

\item{plot.heatmap}{logical, should junction vs trait correlation heatmap plot be saved as output?. Default is FALSE}

\item{test.type}{a character string specifying the type of junction-trait association test to use. Options are "Correlation", "Regression" or "both".}

\item{cor.method}{a character string specifying the method to be used for correlation as in WGCNA. Options are "bicor", "pearson", "kendall" or "spearman".}

\item{useModel}{integer specifying the model type to be used for junction-trait association as in MatrixEQTL. Options are modelLINEAR, modelANOVA or modelLINEAR_CROSS. Valid when test.type = "Regression" or "both".}

\item{covariates}{a numeric vector or a matrix of external sample covariates to be used in regression models. Valid when test.type = "Regression" or "both".}
}
\value{
When "Correlation" is selected as test.type, DJEvsTrait returns the matrices with correlation coefficients and associated P-values for each junction-trait pair.
sig.cor contains top significant associations between junctions in select.junctions and traits for SpliceRadar plotting.
When selected "Regression" as test.type, DJEvsTrait returns a data frame with significant junction-trait associations based on selected model in useModel.
"both" test.type returns correlation and regression outputs.
}
\description{
Correlation between junction expression and sample traits
}
\examples{
data(DJEanlz)
SF <- system.file("extdata", "SF.expr.rds", package = "DJExpress")
SF.exp <- readRDS(SF)
Group1 <- colnames(DJEanlz$v.norm$E)[grep("SRR", colnames(DJEanlz$v.norm$E))]
DT.out <- DJEvsTrait(analize.out = DJEanlz, Group1 = Group1,traitData = SF.exp,
coeff = 0.2,select.junctions = c("chr3:39450216:39452246:1","chr3:39450216:39452244:1",
"chr3:39448256:39449111:1","chr3:39448256:39449093:1"),
test.type = "Correlation", cor.method = "bicor")
}
\seealso{
\code{\link[WGCNA]{cor}}

\code{\link[MatrixEQTL]{Matrix_eQTL_main}}
}
