% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JAn.R
\name{JAn}
\alias{JAn}
\title{JAn: Gene correspondence for splice junctions based on transcriptome annotation}
\usage{
JAn(gtf)
}
\arguments{
\item{gtf}{Reference transcriptome in a gtf file. Used to define gene identity of junctions}
}
\value{
Data frame with junction IDs and Associated gene ID
}
\description{
Identifies splice junctions based on the provided transcriptome annotation (gtf file)
}
\examples{
\dontrun{
gtf <- system.file("extdata", "chr1.gtf.gz", package = "DJExpress")
Jan.out <- JAn(gtf)
}
}
