% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCNAgenePrepare.R
\name{JCNAgenePrepare}
\alias{JCNAgenePrepare}
\title{JCNAGenePrepare: WGCNA annotation for junction expression network construction}
\usage{
JCNAgenePrepare(
  pass1.out,
  genExpr,
  Group1 = NULL,
  gtf,
  nThreads = 2,
  abline.threshold = NULL,
  networkType = "unsigned",
  level = 0.95,
  cor.method = c("bicor", "pearson", "kendall", "spearman")
)
}
\arguments{
\item{pass1.out}{output object from JCNA1pass()}

\item{genExpr}{Gene expression data. A matrix (preferred) or data frame in which columns are genes and rows ar samples. NAs are allowed, but not too many. See WGCNA::checkMissingData and details.}

\item{Group1}{vector or factor specifying basic control sample names}

\item{gtf}{Reference transcriptome in a gtf file. Used to define gene identity of junctions}

\item{nThreads}{numeric: number of threads to allow.}

\item{abline.threshold}{numeric: height cut value in sample dendrogram used to remove offending sample(s)}

\item{networkType}{network type. Allowed values are (unique abbreviations of) "unsigned", "signed", "signed hybrid". See adjacency.}

\item{level}{numeric: confidence interval for Linear Regression of Junction vs Gene trait association value}

\item{cor.method}{a character string specifying the method to be used for correlation as in WGCNA. Options are "bicor", "pearson", "kendall" or "spearman".}
}
\value{
A list object containing:
Gene.sampletree: Sample dendrogram based on gene expression
Gene.hc: Sample dendrogram with trait heatmap
Gene.NetTop: Analysis of network topology for various soft-thresholding powers produced by pickSoftThreshold in WGCNA.
Left panel shows scale-free fit index (y-axis) as a function of the soft-thresholding power (x-axis).
The right panel displays mean connectivity (degree, y-axis) as a function of the soft-thresholding power (x-axis).
Gene.net: Constructed gene network and detected modules using blockwiseModules from WGCNA
Gene.ModuleTrait: Heatmap with Gene Modules vs Trait associations
Genetrait: Data frame with correlation coefficients and associated P-values (Gene expression vs traits)
GeneToJunct: Data frame with junction coordinates and correspondent gene
JunctGeneTrait: List of data frames with integrated information about junction and gene associations per trait
}
\description{
Performs weighted junction co-expression network analysis using gene expression and returns output object for JCNA2pass.
}
\examples{
\dontrun{
gtf0 <- system.file("extdata", "chr1.gtf.gz", package = "DJExpress")
gexp <- system.file("extdata", "genExpr.rds", package = "DJExpress")
filgenExpr <- readRDS(gexp)
J1pass0 <- system.file("extdata", "J1pass.rds", package = "DJExpress")
J1pass <- readRDS(J1pass0)
JgPrep <- JCNAgenePrepare(pass1.out = J1pass, genExpr = filgenExpr,
                          gtf=gtf0, networkType = "unsigned", cor.method = "bicor")
}
}
