% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCNAModTrait.R
\name{JCNAModTrait}
\alias{JCNAModTrait}
\title{JCNAModTrait: Module membership vs. Junction significance plot}
\usage{
JCNAModTrait(
  pass.out,
  nThreads = 2,
  trait,
  module,
  cor.method = c("bicor", "pearson", "kendall", "spearman")
)
}
\arguments{
\item{pass.out}{output object from JCNA1pass() or JCNA2pass()}

\item{nThreads}{numeric: number of threads to allow}

\item{trait}{selected trait for junction module association}

\item{module}{selected junction module for trait association}

\item{cor.method}{a character string specifying the method to be used for correlation as in WGCNA. Options are "bicor", "pearson", "kendall" or "spearman".}
}
\value{
Interactive Highcharts plot with module membership vs junction significance for trait
}
\description{
Plots module membership vs junction significance to identify junctions with high significance for selected trait as well as high module membership in target modules
}
\examples{
J1p <- system.file("extdata", "J1pass.rds", package = "DJExpress")
J1pass <- readRDS(J1p)
jMT <- JCNAModTrait(J1pass, trait ="DDX1",module = "brown", cor.method = "bicor")
}
