% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JCNAprepare.R
\name{JCNAprepare}
\alias{JCNAprepare}
\title{JCNAprepare: WGCNA annotation for junction expression-based network construction}
\usage{
JCNAprepare(
  input.type = c("DJEanalize.out", "junction.counts"),
  workDir = NULL,
  data.type = c("sample", "matrix"),
  aligner = c("STAR", "other"),
  analize.out = NULL,
  Group1 = NULL,
  gtf = NULL,
  minMean = 10,
  maxMean = Inf,
  minVar = 0,
  maxVar = Inf,
  calcNormFactors = FALSE,
  calMethod = "none",
  MVtrend.plot = FALSE,
  traitData = NULL,
  TSdendrogram = FALSE,
  nThreads = 2,
  abline.threshold = NULL
)
}
\arguments{
\item{input.type}{input object type. Options are DJEanalize() output object or junction read counts table.}

\item{workDir}{path to folder where individual junction quantification files (e.g. from STAR alignment) are located. This folder should only contain junction quantification files.}

\item{data.type}{One of c("sample", "matrix"), indicating whether the input files are matrices per individual samples or joint into a single expression matrix. Only used when input.type = "junction.counts".}

\item{aligner}{One of c("STAR", "other"), indicating the alignment tool used to produce junction quantification. Only used if data.type = "matrix".
#' If "other" is indicated, files in workDir path should contain junction IDs (with the format chr:start:end:strand) in the first column read counts in the second column.}

\item{analize.out}{output object from DJEanalize(). Only used if input.type = "DJEanalize.out".}

\item{Group1}{vector or factor specifying basic control sample names. Only used if input.type = "DJEanalize.out".}

\item{gtf}{Reference transcriptome in genecode gtf file format. Used to define gene ID for junctions if input.type = "junction.counts".}

\item{minMean}{numeric: minimum of read count mean per junction. Only used if input.type = "junction.counts".}

\item{maxMean}{numeric: maximum of read count mean per junction. Only used if input.type = "junction.counts".}

\item{minVar}{numeric: minimum of read count variance per junction. Only used if input.type = "junction.counts".}

\item{maxVar}{numeric: maximum of read count variance per junction. Only used if input.type = "junction.counts".}

\item{calcNormFactors}{logical, should edgeR's library Size Normalization be calculated?. Only used if input.type = "junction.counts".}

\item{calMethod}{Normalization method to be used. Options are: "TMM","RLE","upperquartile" or "none" (default "none"). Only used if input.type = "junction.counts".}

\item{MVtrend.plot}{logical, should median/variance trend plot be displayed?. Only used if input.type = "junction.counts".}

\item{traitData}{a numeric vector or a matrix of external sample traits. Samples should be as rows and traits as columns.}

\item{TSdendrogram}{logical, should sample dendrogram plot with trait annotation be displayed?.}

\item{nThreads}{numeric: number of threads to allow.}

\item{abline.threshold}{numeric: height cut value in sample dendrogram used to remove offending sample(s).}
}
\value{
List object containing sample dendrogram, expression data matrix, trait data matrix and topology network plots
}
\description{
Filters samples and junctions in DJEanalize expression data object for junction co-expression network analysis.
}
\examples{
\dontrun{
data(DJEanlz)
SF <- system.file("extdata", "SF.expr.rds", package = "DJExpress")
SF.exp <- readRDS(SF)
colnames(DJEanlz$v.norm)[grep("TCGA", colnames(DJEanlz$v.norm))] <- paste0("TCGA_",
seq(1,length(colnames(DJEanlz$v.norm)[grep("TCGA", colnames(DJEanlz$v.norm))]), 1))
Group1 <- colnames(DJEanlz$v.norm$E)[grep("SRR", colnames(DJEanlz$v.norm$E))]
Jprep <- JCNAprepare(analize.out=DJEanlz, Group1 = Group1,
traitData = SF.exp, abline.threshold=60, input.type = "DJEanalize.out")
}
}
