% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topDJE.R
\name{topDJE}
\alias{topDJE}
\title{topDJE: Top table of differentially used junctions}
\usage{
topDJE(
  fit,
  coef = ncol(fit),
  test = "simes",
  number = 10L,
  FDR = 1,
  sort.by = "p"
)
}
\arguments{
\item{fit}{MArrayLM fit object produced by dje}

\item{coef}{the coefficient (column) of fit for which differentially splicing is assessed.}

\item{test}{character string, possible values are "simes", "F" or "t". "F" gives F-tests for each gene. "t" gives t-tests for each exon. "simes" gives genewise p-values derived from the t-tests after Simes adjustment for each gene.}

\item{number}{integer, maximum number of rows to output.}

\item{FDR}{numeric, only show junctions or genes with false discovery rate less than this cutoff.}

\item{sort.by}{character string specifying which column to sort results by. Possible values for "p", "logFC", "Njunctions" or "none". "logFC" is only available if test="t" and "Njunctions" is only available if test="simes" or test="F".}
}
\value{
A data.frame with any annotation columns found in fit plus the following columns as in limma::topSplice
}
\description{
Top table ranking the most differentially spliced junctions. This is a wrapper function of limma topSplice.
}
\examples{
\dontrun{
fit.norm0 <- system.file("extdata", "dje.out.rds", package = "DJExpress")
fit.norm <- readRDS(fit.norm0)
topdje.out <- topDJE(fit = fit.norm)
}
}
\seealso{
\code{\link[limma]{topSplice}}
}
